% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateKdXd.R
\name{calculateKdXd}
\alias{calculateKdXd}
\title{calculateKdXd function}
\usage{
calculateKdXd(geno_data, pheno_data, D, site_qtl_dom, m_qtl_dom)
}
\arguments{
\item{geno_data}{genotype data}

\item{pheno_data}{phenotype data}

\item{D}{dominance relationship matrix}

\item{site_qtl_dom}{a vector, names of dom QTL sites}

\item{m_qtl_dom}{the number of dom QTLs}
}
\value{
list(Kd = Kd,Xd = Xd)
}
\description{
calculateKdXd function
}
\examples{
\dontrun{KdXd = calculateKdXd(mmgeno_data, mmpheno_data, D, site_qtl, m_qtl)}
}
