% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno.generate.R
\name{pheno.generate}
\alias{pheno.generate}
\title{pheno.generate function}
\usage{
pheno.generate(model, geno_data, effects, sigma.error)
}
\arguments{
\item{model}{a character that indicate the genetic effects included ("A" or "AD")}

\item{geno_data}{geno_data given from geno.generate() function}

\item{effects}{an M*p matrix p is the number of genetic effects included (such as "A" for 1, "AD" for 2)}

\item{sigma.error}{the variance for residual effects.}
}
\value{
pheno_data a data frame, with two columns $GID, and $SimTrait
}
\description{
pheno.generate function
}
\examples{
\dontrun{pheno.generate(genotypes = t(geno_data[-c(1:3)]), effects = b,
                   indNum = indNum, sigma.error = sigma_error)}
}
