% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtxnetwork.output.trans.R
\name{qtxnetwork.output.trans}
\alias{qtxnetwork.output.trans}
\title{qtxnetwork.output.trans}
\usage{
qtxnetwork.output.trans(pheno_data, pre_file)
}
\arguments{
\item{pheno_data}{a data frame of phenotypic data, one of the outputs from simulation}

\item{pre_file}{path and name for .pre file}
}
\value{
a list
        $ qtl_data a data frame with additive qtl information
        $ qtl_dom_data a data frame with dominance qtl information
}
\description{
qtxnetwork.output.trans
}
\examples{
\dontrun{qtl = qtxnetwork.output.trans(pheno_data, pre_file)
                   qtl$qtl_data
                   qtl$qtl_dom_data}
}
