% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp.effect.R
\name{snp.effect}
\alias{snp.effect}
\title{snp.effect function}
\usage{
snp.effect(
  model,
  snpNum,
  major_a_idx,
  variance_a_major,
  variance_a_minor,
  major_d_idx,
  variance_d_major,
  variance_d_minor
)
}
\arguments{
\item{model}{a character that indicate the genetic effects included ("A" or "AD")}

\item{snpNum}{The number of SNPs}

\item{major_a_idx}{The index for major additive SNP}

\item{variance_a_major}{The variance for major additive SNP effect}

\item{variance_a_minor}{The variance for minor additive SNP effect}

\item{major_d_idx}{The index for major dominance SNP}

\item{variance_d_major}{The variance for major dominance SNP effect}

\item{variance_d_minor}{The variance for minor dominance SNP effect}
}
\value{
effect an M*p matrix p is the number of genetic effects included (such as "A" for 1, "AD" for 2)
}
\description{
snp.effect function
}
\examples{
snp.effect(model = "AD", snpNum = 2000,
                     major_a_idx = c(500, 750, 1000, 1250, 1500), variance_a_major = 0.02, variance_a_minor = 0.002,
                     major_d_idx = c(500, 750, 1000, 1250, 1500), variance_d_major = 0.02, variance_d_minor = 0.002)
}
