% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmdata.R
\name{mmdata}
\alias{mmdata}
\title{mmdata function
Format data for mmGEBLUP analysis}
\usage{
mmdata(geno_data, pheno_data, qtl_data, qtl_env_data)
}
\arguments{
\item{geno_data}{input genotype data}

\item{pheno_data}{input phenotype data}

\item{qtl_data}{input additice qtl data}

\item{qtl_dom_data}{input dominance qtl data}
}
\value{
a list with QCed genotype and phenotype data, and other data required for GS analysis
        $ mmgeno_data an n*m matrix
        $ mmpheno_data an n_obs*3 data frame
        $ Ka additive relationship matrix for mmGBLUP
        $ Xa additive fixed effect coefficient for mmGBLUP
        $ Kd dominance relationship matrix for mmGBLUP
        $ Xd dominance fixed effect coefficient for mmGBLUP
        $ A additive relationship matrix for GBLUP
        $ D additive relatinoship matrix for GBLUP
        $ mmsummary data summary
}
\description{
mmdata function
Format data for mmGEBLUP analysis
}
\examples{
\dontrun{qcdata = dataqc(geno_data, pheno_data, qtl_data, qtl_env_data)}
}
