% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmgblup.R
\name{mmgblup}
\alias{mmgblup}
\title{mmgblup function}
\usage{
mmgblup(data, Ka, Kd)
}
\arguments{
\item{data}{a data frame}

\item{Ka}{additive genetic relationship matrix}

\item{Kd}{dominance genetic relationship matrix}
}
\value{
list(mod, BV)
}
\description{
mmgblup function
}
\examples{
\dontrun{rst = mmgblup(data = cbind(dt, mmdata$Xa, mmdata$Xd), Ka = mmdata$Ka, Kd = mmdata$Ka)}
}
